Consolidate modals for edit and create svc type
